Remove suffix from submodule url
I have to set my remotes to github.com-joyfulnoisesynth so I can
work with the SSH key for that account, but if I leave that suffix
in the submodule url, GitHub doesn't realize the submodule comes
from another GitHub repo.

Luckily, it seems like the main repo's configuration passes through
to the submodules, so I don't even need the suffix there!