[JNTUB-HW] Create SMD version of PCB
Use KiCad scripting API to write a script to turn most
through-hole parts into SMT parts. Will use this to release
two variants of the board.

Had to reroute many traces so that every through hole pad is
connected to traces on the front layer (that way, when all
the through holes are removed, the components are still
connected to traces).